增加异常处理取代错误码，但是现在的实现方案不是特别优雅，等后续熟悉的框架的依赖注入之后再做修改，但是修改不会影响使用。大家可以采用project接口中的异常处理，也可以仍然采用错误码的方法

+++ app/plugins/SecurityPlugin.php
@@ -93,7 +93,7 @@ class SecurityPlugin extends Plugin
 	 * @param Dispatcher $dispatcher
 	 */
 	public function beforeDispatch(Event $event, Dispatcher $dispatcher)
-	{
+	{	
 		$auth = $this->session->get('auth');
 		if (!$auth){
 			$role = 'Guests';
@@ -106,12 +106,16 @@ class SecurityPlugin extends Plugin
 
 	 	$acl = $this->getAcl();
 	 	$allowed = $acl->isAllowed($role, $controller, $action);
+	 	
 		if ($allowed != Acl::ALLOW) {
-			$dispatcher->forward(array(
-				'controller' => 'errors',
-				'action'     => 'show401'
-			));
-            $this->session->destroy();
+			// $dispatcher->forward(array(
+			// 	'controller' => 'errors',
+			// 	'action'     => 'show401'
+			// ));
+
+			if (empty($this->session)) {
+				$this->session->destroy();
+			}
 			return false;
 		}
 	}
